Fixed login/logout enchantment bugs
Fixed bug where logging out and back in did not remove the bonus health
from armor enchants. This meant that logging out and back in kept giving
you extra health = to the armor health bonus. Thus Bonus is removed on
logout.

Added client and server side health updating when logging in with armor
(player health is set to loggedOutHealth on client and server side now)